impr: package managers handling
