+++ src/managers/exec.ts
@@ -1,0 +1,41 @@
+import * as cp from "node:child_process";
+import { CancellationToken } from "../util";
+
+interface IOptions {
+	cwd?: string;
+	stdio?: any;
+	customFds?: any;
+	env?: any;
+	timeout?: number;
+	maxBuffer?: number;
+	killSignal?: string;
+}
+
+export function exec(
+	command: string,
+	options: IOptions = {},
+	cancellationToken?: CancellationToken
+): Promise<{ stdout: string; stderr: string }> {
+	return new Promise((c, e) => {
+		let disposeCancellationListener: Function | null = null;
+
+		const child = cp.exec(command, { ...options, encoding: 'utf8' } as any, (err, stdout: string, stderr: string) => {
+			if (disposeCancellationListener) {
+				disposeCancellationListener();
+				disposeCancellationListener = null;
+			}
+
+			if (err) {
+				return e(err);
+			}
+			c({ stdout, stderr });
+		});
+
+		if (cancellationToken) {
+			disposeCancellationListener = cancellationToken.subscribe((err: any) => {
+				child.kill();
+				e(err);
+			});
+		}
+	});
+}
