+++ src/managers/none.ts
@@ -1,0 +1,13 @@
+import type { IPackageManager } from "./manager";
+import { pmNPM } from './npm';
+
+export const pmNone: IPackageManager = {
+	binaryName: "",
+	selfVersion: pmNPM.selfVersion.bind(pmNPM),
+	selfCheck: pmNPM.selfCheck.bind(pmNPM),
+	pmRunCommand: pmNPM.pmRunCommand,
+	async pmProdDependencies(cwd: string, _?: string[]): Promise<string[]> {
+		return [cwd]
+	},
+	pmFetchLatestVersion: pmNPM.pmFetchLatestVersion.bind(pmNPM),
+}
